+++ bluetooth_low_energy_platform_interface/lib/src/bluetooth_low_energy_manager.dart
@@ -33,4 +33,17 @@ abstract base class BaseBluetoothLowEnergyManager extends PlatformInterface
 
   /// Initializes the [BaseBluetoothLowEnergyManager].
   void initialize();
+
+  /// Requests permissions to be granted to this application. These permissions
+  /// must be requested in your manifest, they should not be granted to your app,
+  /// and they should have protection level dangerous, regardless whether they
+  /// are declared by the platform or a third-party app.
+  ///
+  /// This method is available on Android, throws [UnsupportedError] on other platforms.
+  Future<bool> authorize();
+
+  /// Show screen of details about a particular application.
+  ///
+  /// This method is available on Android, throws [UnsupportedError] on other platforms.
+  Future<void> showAppSettings();
 }